ref(performance): Extract shared trace issue utils into dedicated module
Deduplicates six functions, one interface, and five constants that were
copy-pasted between traceIcons.tsx and virtualizedViewManager.tsx.
